requirements.txt
beautifulsoup4==4.8.0
peewee==3.13.2
requests==2.21.0
python-dotenv==0.12.0
psycopg2-binary==2.8.5
py-postgresql==1.2.1